Draw data labels on the bottom and remove value labels from buckets

+++ Source/Charts/Charts/CandleStickChartView.swift
@@ -32,5 +32,5 @@ open class CandleStickChartView: BarLineChartViewBase, CandleChartDataProvider
         return _data as? CandleChartData
     }
     
-    open var positions = [CGFloat]()
+    open var positions = [Float]()
 }
