+++ Source/Charts/Charts/HighLowChartView.swift
@@ -28,5 +28,5 @@ open class HighLowChartView: BarLineChartViewBase, CandleChartDataProvider
         return _data as? CandleChartData
     }
     
-    open var positions = [CGFloat]()
+    open var positions = [Float]()
 }
